ai/pyproject.toml
[project]
name = "crashwise"
version = "0.7.3"
description = "Crashwise AI orchestration module"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk",
    "a2a-sdk",
    "litellm",
    "python-dotenv",
    "httpx",
    "uvicorn",
    "rich",
    "agentops",
    "fastmcp",
    "mcp",
    "typing-extensions",
    "cognee>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "black",
    "ruff",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/Crashwise"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.uv]
dev-dependencies = [
    "pytest",
    "pytest-asyncio",
]
